Add UnitTests for GenerateGithubWebhook
